backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
langchain==0.1.0
langchain-openai==0.0.2
faiss-cpu
sentence-transformers==2.2.2
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.0
arxiv==2.1.0
